requirements.txt
flask
speechmatics-python
httpx
gunicorn 